pyproject.toml
[project]
name = "sensor-data-sim"
version = "0.2.0"
description = "Project centered around the generation, processing, analysis, and display of mock sensor data."
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[tool.uv.workspace]
members = ["data/generators", "data/writers", "data/transformers", "data/streams"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselct with '-m \"not slow\"')",
    "veryslow: marks tests as very slow (deselect with '-m \"not veryslow\"')",
]

[dependency-groups]
test = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "moto[s3]>=5.1.4",
    "pytest-mock>=3.14.0",
]
lint = [
    "basedpyright>=1.28.4",
    "boto3-stubs[s3]>=1.38.2",
    "pandas-stubs>=2.2.3.250308",
    "pyarrow-stubs>=19.1",
    "ruff>=0.11.6",
    "scipy-stubs>=1.15.2.2",
]
docs = [
    "pydata-sphinx-theme>=0.16.1",
    "sphinx>=8.2.3",
]
dev = [
    "jupyterlab>=4.4.1",
    "matplotlib>=3.10.1",
]
